+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+* Fixing issue with original ratio not being saved.
+* New method `pickImagesAsByData` to pick images as Uint8List.
+
 ## 0.1.0
 
 * You can now use `width` and `height` arguments to set the size of the image.